refactor: `favorite_view` for Streamlined Product Liking
This commit refactors the `favorite_view` to streamline the process of liking products and improve the overall user experience. The primary changes to the view include:

1. Transition to JavaScript-Based Liking:
The `favorite_view` has been adapted to work in conjunction with the JavaScript-based liking system (implemented in `like_product.js`).

2. Enhanced Error Handling:
Robust error handling has been incorporated to provide users with clearer and more user-friendly error messages, especially in cases like JSON decoding errors.

3. HTTP Method Restriction:
The view is now decorated with `@require_POST` to ensure that liking products can only be performed via POST requests. This helps maintain proper data integrity and security.

4. Authentication Validation:
The view now thoroughly checks the user's authentication status before processing like requests. If the user is not authenticated, they are appropriately redirected to the login page.

5. Responses:
Two distinct `JsonResponse` instances have been implemented to ensure precise communication between the frontend and backend.
    - When the user is authenticated, the view returns a JSON response with `{'authenticated': True}`.
    - When the user is not authenticated, the response includes `{'authenticated': False, 'login': request.build_absolute_uri(reverse('account_login'))}`, providing the login URL for use in `like_product.js`. This URL is used to redirect the user to the login page when necessary.

By implementing these changes, the `favorite_view` aligns with the JavaScript-driven liking system, ensuring a smooth and consistent liking experience while maintaining reliable error handling and authentication checks.

refactor: Simplify URL for Product Liking

This commit simplifies the URL pattern for product liking functionality by removing the `<int:pk>` parameter. In the corresponding view, product identification is now handled through JSON data extracted from the request body. This change eliminates unnecessary complexity in the URL structure, making it more straightforward and aligning it with the way product identification is managed in the view.

refactor: Introduce 'like-product' Buttons for Improved User Experience

In this commit, anchor tags (<a>) in the product_list_base.html template are replaced with buttons (<button>) for product liking. These buttons now feature the like-product class, aligning them with the functionality provided by the like_product.js script. This enhancement ensures a consistent design and makes liking products more user-friendly.

+++ products/views.py
@@ -1,11 +1,14 @@
-from django.shortcuts import Http404, get_object_or_404, redirect, render
+from django.shortcuts import Http404, get_object_or_404, redirect, render, reverse
 from django.views import generic
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_POST
 from django.utils.decorators import method_decorator
 from django.contrib import messages
 from django.utils.translation import gettext_lazy as _
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, JsonResponse
 from django.db.models import Q, F, Case, When, IntegerField
+
+import json
 
 from .models import Product, TimeLike, ProductComment, Category
 from .forms import ProductCommentForm, SearchForm
@@ -124,21 +127,34 @@ class SearchView(generic.ListView):
         return super().dispatch(request, *args, **kwargs)
 
 
-@login_required
-def favorite_view(request, pk):
-    product_obj = get_object_or_404(Product, pk=pk, active=True)
-    user = request.user
-    if product_obj.favorite.filter(pk=user.pk).exists():
-        product_obj.favorite.remove(user)
-        get_object_or_404(TimeLike, user=user, product=product_obj).delete()
-        messages.error(request, _('Unlike post.'))
-    else:
-        product_obj.favorite.add(user)
-        time_like, create = TimeLike.objects.get_or_create(user=user, product=product_obj)
-        time_like.save()
-        messages.success(request, _('Like post.'))
+@require_POST
+def favorite_view(request):
+    try:
+        data = json.loads(request.body)
+    except json.JSONDecodeError:
+        messages.warning(request, _('Oops! Something went wrong with your request. Please try again.'
+                                    ' If the issue persists, contact our support team for assistance.'))
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
 
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    if request.user.is_authenticated:
+        pk = data.get('productId')
+        product_obj = get_object_or_404(Product, pk=pk, active=True)
+        user = request.user
+        if product_obj.favorite.filter(pk=user.pk).exists():
+            product_obj.favorite.remove(user)
+            get_object_or_404(TimeLike, user=user, product=product_obj).delete()
+            messages.error(request, _('Unlike post.'))
+        else:
+            product_obj.favorite.add(user)
+            time_like, create = TimeLike.objects.get_or_create(user=user, product=product_obj)
+            time_like.save()
+            messages.success(request, _('Like post.'))
+
+        response = {'authenticated': True}
+        return JsonResponse(response, safe=False)
+    else:
+        response = {'authenticated': False, 'login': request.build_absolute_uri(reverse('account_login'))}
+        return JsonResponse(response, safe=False)
 
 
 class CategoryView(generic.ListView):
